Больше противников

+++ core/src/com/warfare/darkannihilation/systemd/service/Windows.java
@@ -6,8 +6,8 @@ public final class Windows {
     public static final int SCREEN_HEIGHT = 1080;
     public static final int HALF_SCREEN_HEIGHT = 540;
 
-    public static float HALF_SCREEN_WIDTH;
-    public static float SCREEN_WIDTH;
+    public static int HALF_SCREEN_WIDTH;
+    public static int SCREEN_WIDTH;
 
     public static int PHONE_WIDTH;
     public static int PHONE_HEIGHT;
@@ -16,7 +16,7 @@ public final class Windows {
         PHONE_WIDTH = Gdx.graphics.getWidth();
         PHONE_HEIGHT = Gdx.graphics.getHeight();
 
-        SCREEN_WIDTH = PHONE_WIDTH / (PHONE_HEIGHT / (float) SCREEN_HEIGHT);
-        HALF_SCREEN_WIDTH = SCREEN_WIDTH / 2f;
+        SCREEN_WIDTH = (int) (PHONE_WIDTH / (PHONE_HEIGHT / (float) SCREEN_HEIGHT));
+        HALF_SCREEN_WIDTH = SCREEN_WIDTH / 2;
     }
 }
